[GCC] Fix issues reported by GCC (#5009)
These are benign issues that should silence some warnings.

MODULAR_ORIG_COMMIT_REV_ID: 77258cdb80eddd8545406992072bccaf6400f35f

+++ KGEN/lib/KGENToLLVM/LowerPOPToLLVM.cpp
@@ -800,7 +800,8 @@ struct ConvertPOPStructGEP : mlir::ConvertOpToLLVMPattern<POP::StructGEPOp> {
       return op.emitError("failed to convert result type");
     rewriter.replaceOpWithNewOp<LLVM::GEPOp>(
         op, ptrType, adaptor.getContainer(),
-        ArrayRef<LLVM::GEPArg>{0, op.getIndexAttr().getInt()});
+        ArrayRef<LLVM::GEPArg>{
+            0, static_cast<int32_t>(op.getIndexAttr().getInt())});
     return success();
   }
 };
